+++ src/components/common/TextField/index.tsx
@@ -1,2 +1,2 @@
 export { default as TextField } from "./TextField";
-export { StyledWrapper, StyledInput } from "./TextFiled.Styled";
+export { StyledWrapper, StyledInput } from "./TextField.Styled";
